% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_tran.R
\name{bbox_tran}
\alias{bbox_tran}
\title{Generate bounding box for Neotoma}
\usage{
bbox_tran(x, coord_formula = "~ x + y", from, to)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix} with the vegetation data and its coordinates.}

\item{from}{The proj4 string.}

\item{formula}{The formula, as a string}
}
\description{
From a vegetation matrix with \code{x} and \code{y} columns (or \code{lat}/\code{long}), generate a bounding box to be used for the \code{loc} parameter in the \code{neotoma} function \code{get_dataset()}.
}
