% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R
\name{build_grid}
\alias{build_grid}
\title{Create the raster grid for the analysis.}
\usage{
build_grid(veg_box, resolution = 8000, proj = "+init=epsg:3175")
}
\arguments{
\item{veg_box}{The bounding box for the vegetation data, as a vector with elements \code{c(W,S,N,E)}.}

\item{resolution}{The grid resolution for the output raster, in the native output projection.}

\item{proj}{A projection string using \code{proj4} standards.  Default is Gt. Lakes & St. Lawrence Albers: \code{'+init=epsg:3175'}.}
}
\value{
A raster, returned from the \code{raster} package.
}
\description{
Create the raster grid for the analysis.
}
\examples{
  \dontrun{
   veg_mean <- readr::read_csv('data/composition_v0.3.csv')
   veg_box <- bbox_tran(veg_mean, '~ x + y',
                        '+init=epsg:3175',
                        '+init=epsg:3175')
   reconst_grid <- build_grid(veg_box,
                              resolution = 8000,
                              proj = '+init=epsg:3175')
  }

}
