% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\docType{data}
\name{downloads}
\alias{downloads}
\title{Neotoma download data for the Upper Midwestern United States}
\format{An object of class \code{download_list} (inherits from \code{list}) of length 348.}
\source{
\url{http://neotomadb.org}
}
\usage{
downloads
}
\description{
A set of \code{download} objects from the Neotoma Paleoecological Database
to be used in the STEPPS package vignette, or as part of code examples.
}
\author{
Simon Goring
}
\keyword{datasets}
