% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Title}
\usage{
generate_tables(taxa, output = NULL)
}
\arguments{
\item{taxa}{A table containing taxa, either from a `download`, a `download_list` or}

\item{output}{A string representing the file path for export.  Can be NULL.}
}
\value{
A \code{data.frame}.
}
\description{
Title
}
\examples{
{
}

}
