% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_input.R
\name{prep_input}
\alias{prep_input}
\title{Preparing output for the STAN run of the STEPPS Model}
\usage{
prep_input(veg, pollen, target_taxa, grid, hood = 7e+05, dist_scale = 1e+06)
}
\arguments{
\item{veg}{A table of vegetation data, generally gridded.}

\item{pollen}{A table of available pollen data.}

\item{target_taxa}{The taxa to be reconstructed.  Must appear in both the vegetation and pollen data.}
}
\value{
A list with parameters:
\itemize{
\item K  - Number of taxa
\item N_cores - Number of cores
\item N_cells - Number of cells in the veg data
\item N_hood - Number of veg cells contributing to each pollen sample site
\item y - Pollen counts (\emph{n} cores x \emph{p} taxa - a wide table)
\item r - Veg proportions (same structure)
\item idx_cores - The index of the veg cell for each core
\item idx_hood - The indices of cells for each contributing neighborhood
\item d - distance matrix (spatial distance from cores to veg cells)
\item N_pot - Number of potential contributing cells
\item d_pot - The actual contributing cells for each pollen sample
}
}
\description{
Preparing output for the STAN run of the STEPPS Model
}
\examples{
{
   stepps_input <- prep_input(veg    = veg_table,
                              pollen = pol_table,
                              target_taxa = target_taxa,
                              grid   = reconst_grid)
}
}
