% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_timebins.R
\name{pull_timebins}
\alias{pull_timebins}
\title{Get calibration pollen samples by time.}
\usage{
pull_timebins(x, calib_range = c(200, 400))
}
\arguments{
\item{x}{A \code{download} or \code{download_list} object.}

\item{calib_range}{The age ranges to extract.}
}
\value{
a \code{data.frame}
}
\description{
With a given set of bounding dates, extract a calibration dataset from a `download` or `download_list` object frmo `neotoma`.
}
