% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_taxa.R
\name{translate_taxa}
\alias{translate_taxa}
\title{Translate to target taxonomy.}
\usage{
translate_taxa(input_table, translation, id_cols)
}
\arguments{
\item{input_table}{The composition data table, consisting of \code{id_var} columns, indicating spatial and compositional data, with one taxon in each column.}

\item{translation}{The translation table, generated by \code{generate_table}.}

\item{id_cols}{The location columns, for example \code{c('x', 'y')}, representing the spatial coordinates for the data.}
}
\value{
A \code{data.frame} with indicator columns as identified in \code{id_cols} and once column for each of the unique taxa.
}
\description{
Using a translation table as generated by \code{generate_table} and modified by the user,
translate the target composition data to the new taxonomic groupings.
}
\examples{
{
calib_trans <- translate_taxa(calib_dialect,
                              pol_table,
                              id_cols = colnames(calib_dialect)[1:10])
}
}
