% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grid.R
\name{build_grid}
\alias{build_grid}
\title{Create the raster grid for the analysis.}
\usage{
build_grid(veg_box, resolution = 8000, proj = "+init=epsg:3175")
}
\arguments{
\item{veg_box}{The bounding box for the vegetation data, as a vector with elements \code{c(W,S,N,E)}.}

\item{proj}{}
}
\description{
Create the raster grid for the analysis.
}
