% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tables.R
\name{generate_tables}
\alias{generate_tables}
\title{Generate tables}
\usage{
generate_tables(taxa, output = NULL)
}
\arguments{
\item{taxa}{A table containing taxa, either from a `download`, a `download_list` or}

\item{output}{A string representing the file path for export.  Can be NULL.}
}
\value{
A \code{data.frame} with two columns: \describe{
  \item{target}{The original pollen taxon to be translated}
  \item{match}{The simplified taxon to be used in calibration}
}
}
\description{
This function is used to help support the generation of the taxon
translation tables, by exporting a \code{download} or \code{download_list} file's
taxon table.
}
\examples{
{
data(downloads)
generate_tables(downloads, output = 'data/translation.csv')
}

}
