% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_input.R
\name{prep_input}
\alias{prep_input}
\title{Preparing output for the STAN run of the STEPPS Model}
\usage{
prep_input(veg, pollen, target_taxa, grid, hood)
}
\arguments{
\item{target_taxa}{}
}
\value{
A list with parameters:
\itemize{
\item K  - Number of taxa
\item N_cores - Number of cores
\item N_cells - Number of cells in the veg data
\item N_hood - Number of veg cells contributing to each pollen sample site
\item y - Pollen counts (\emph{n} cores x \emph{p} taxa - a wide table)
\item r - Veg proportions (same structure)
\item idx_cores - The index of the veg cell for each core
\item idx_hood - The indices of cells for each contributing neighborhood
\item d - distance matrix (spatial distance from cores to veg cells)
\item N_pot - Number of potential contributing cells
\item d_pot - The actual contributing cells for each pollen sample
}
}
\description{
Preparing output for the STAN run of the STEPPS Model
}
