% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_input.R
\name{prep_input}
\alias{prep_input}
\title{Preparing output for the STAN run of the STEPPS Model}
\usage{
prep_input(veg, pollen, target_taxa, grid, hood = 7e+05, dist_scale = 1e+06)
}
\arguments{
\item{veg}{A table of vegetation data, generally gridded.}

\item{pollen}{A table of available pollen data.}

\item{target_taxa}{The taxa to be reconstructed.  Must appear in both the vegetation and pollen data.}

\item{grid}{A data grid to be used for calibration of class \code{raster}.}

\item{hood}{Distance (in units native to the grid) from which to estimate regional neighborhood for the pollen-vegetation calibration.  Default is \code{1e5}, assumed to be meters (100km).}

\item{dist_scale}{Currently not implemented.}
}
\value{
A list with parameters: \describe{
\item{K}{Number of taxa in the data.}
\item{N_cores}{Number of sedimentary cores used in the analysis.}
\item{N_cells}{Number of cells in the parent grid.}
\item{N_hood}{Number of vegetation cells contributing to each pollen sample site.}
\item{y}{Pollen counts, a matrix with \code{N_cores} rows and \code{K} columns.}
\item{r}{Vegetation proportions, a matrix with \code{N_cells} rows and \code{K} columns.}
\item{idx_cores}{The index for each sedimentary core, with respect to the vegetation grid.}
\item{idx_hood}{The index of each neighbourhood cell for each core, with respect to the vegetation grid.}
\item{d}{Distance matrix from each sedimentary core to each vegetation cell.}
\item{N_pot}{The number of potentially contributing cells for each vegetation cell.}
\item{d_pot}{The actual number of vegetation cells contributing to each pollen sample site.}
}
}
\description{
Preparing output for the STAN run of the STEPPS Model
}
\examples{
{
   stepps_input <- prep_input(veg    = veg_table,
                              pollen = pol_table,
                              target_taxa = target_taxa,
                              grid   = reconst_grid)
}
}
