% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_stepps_shape.R
\name{to_stepps_shape}
\alias{to_stepps_shape}
\title{Convert object to shape.}
\usage{
to_stepps_shape(x, coord_fun, proj)
}
\arguments{
\item{x}{The data.frame to convert to a SpatialPointsDataFrame object}

\item{coord_fun}{The coordinate columns, as a string formula}

\item{proj}{The data ection in `proj` format.}
}
\value{
A \code{SpatialPointsDataFrame}.
}
\description{
Convert object to shape.
}
